Added extra configuration to the plugin to make it highly flexible
