+++ depclean-gradle-plugin/src/main/java/se/kth/depclean/analysis/GradleProjectDependencyAnalyzer.java
@@ -1,16 +1,14 @@
 package se.kth.depclean.analysis;
 
 import org.gradle.api.Project;
-
 import se.kth.depclean.core.analysis.ProjectDependencyAnalyzerException;
 
 public interface GradleProjectDependencyAnalyzer {
-  // fields -----------------------------------------------------------------
 
+  // fields -----------------------------------------------------------------
   String ROLE = GradleProjectDependencyAnalyzer.class.getName();
 
   // public methods ---------------------------------------------------------
-
   GradleProjectDependencyAnalysis analyze(Project project) throws ProjectDependencyAnalyzerException;
 }
