+++ depclean-gradle-plugin/src/main/resources/META-INF/gradle-plugins/se.kth.castor.depCleanGradlePlugin
@@ -1,1 +1,1 @@
-implementation-class = se.kth.depclean.depcleanGradlePlugin
+implementation-class = se.kth.depclean.DepCleanGradlePlugin
