--- depclean-gradle-plugin/src/test/java/sampleTest.java
@@ -1,4 +1,0 @@
-public class sampleTest {
-    int x = 5;
-
-}
