+++ depclean-gradle-plugin/src/test/resources/all_dependencies_unused/src/main/java/Main.java
@@ -1,0 +1,9 @@
+// import com.fasterxml.jackson.databind.ObjectMapper;
+//import com.fasterxml.jackson.core.util.Separators;
+
+
+public class Main {
+   // private static final ObjectMapper converter = new ObjectMapper();
+//   Separators separators = new Separators();
+   int field = 42;
+}
