--- depclean-gradle-plugin/src/test/resources/all_dependencies_unused/src/main/java/test.java
@@ -1,5 +1,0 @@
-public class test {
-    public static void main(String[] args) {
-        System.out.println("Test successful");
-    }
-}
